Весь функционал сегодняшнего дня реализован.

+++ app/src/main/java/ru/bscmsk/renttable/presentation/adapters/VerticalSpaceItemDecoration.kt
@@ -5,12 +5,12 @@ import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.ItemDecoration
 
-class VerticalSpaceItemDecoration(private val VerticalSpaceHeight: Int) :
+class VerticalSpaceItemDecoration(private val VerticalSpaceHeight: Float) :
     ItemDecoration() {
     override fun getItemOffsets(
         outRect: Rect, view: View, parent: RecyclerView,
         state: RecyclerView.State
     ) {
-        outRect.bottom = VerticalSpaceHeight
+        outRect.bottom = VerticalSpaceHeight.toInt()
     }
 }
